+++ docs/reference/extensions.md
@@ -1,0 +1,1 @@
+::: substrateinterface.extensions
