+++ docs/reference/interfaces.md
@@ -1,0 +1,1 @@
+::: substrateinterface.interfaces
