+++ docs/reference/keypair.md
@@ -1,0 +1,1 @@
+::: substrateinterface.keypair
